refactor(toast): move toast utilitary service to utils folder
Also updates its references
