+++ src/app/app.module.ts
@@ -22,7 +22,7 @@ import { environment } from './../environments/environment';
 import { MyApp } from './app.component';
 
 import { StorageProvider } from './../providers/storage/storage';
-import { ToastProvider } from './../providers/toast/toast';
+import { ToastProvider } from './../providers/util/toast/toast';
 import { BitpointDataProvider } from './../providers/bitpoint-data/bitpoint-data.provider';
 import { ClipboardProvider } from './../providers/clipboard/clipboard.provider';
 import { AuthProvider } from './../providers/auth/auth';
